requirements.txt
# Core Libraries
python-telegram-bot==21.4
python-dotenv
llama-cpp-python

psycopg2-binary
pymongo
